requirements.txt
Django==4.2.7
djangorestframework==3.14.0
openai==1.100.2
python-dotenv==1.0.0
whisper-openai==1.0.0
PyMuPDF==1.24,<1.26
chromadb==0.4.15
sentence-transformers==2.7.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
tenacity==8.2.3
